Day Two Commit
Worked a couple of hours on it. Got a handle on control structures,
functions, and the basics of iterators. Didn't get a stateless List
iterator to work, but I don't have enough experience to troubleshoot why
yet. I'll move forward and come back to it as it wasn't actually part of
the tutorial just a side exercise I gave myself.